Add 'smpte' command for testing
Future ideas would be to control the sine wave hertz, volume
(important), and some other things.

Coming up I want to improve the window size / starting position and add
certain arguments for that purpose. Especially when accessing from the
NUC because with that little screen it should always be fullscreen.