Add NES CPU on back side and initial edge cut
More space could be available to the left, but want to avoid RF
shielding if at all possible. This is because when you remove the RF
shielding and don't put it back - you're probably going to lose it and
that is essentially modifying the console.

To the right there are pins from a capacitor that should be left alone.